requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asarPy==1.0.1
attrs==25.4.0
audioop-lts==0.2.2
cffi==2.0.0
Cython==3.2.1
discord.py==2.6.4
discord-ext-voice-recv>=0.1.4
deepl>=1.17.0
frozenlist==1.8.0
googletrans==4.0.0rc1
idna==3.11
Mako==1.3.10.dev0
Markdown==3.10
MarkupSafe==3.0.3
meson==1.9.1
multidict==6.7.0
numpy>=1.24.0
openai>=1.0.0
packaging==25.0
pre-commit>=3.7.0
propcache==0.4.1
pycparser==2.23
PyNaCl==1.6.1
python-dotenv==1.2.1
ruff>=0.12.0
setuptools==80.9.0
wheel==0.45.1
yarl==1.22.0
